Move fips functionality to FIPSSwitch class
